aprobacion de actualizaciones

+++ app/Http/Controllers/CarManagement/CarsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\FleetCar\FleetCarChangelogRequest;
 use App\Http\Requests\FleetCar\FleetCarDocumentRequest;
 use App\Http\Requests\FleetCar\FleetCarRequest;
+use App\Models\ApprovalCarDocument;
 use App\Models\Car;
 use App\Models\CarChangelog;
 use App\Models\CarChangelogItem;
@@ -24,7 +25,7 @@ class CarsController extends Controller
     public function index()
     {
         $user = Auth::user();
-        $cars = Car::with(['user', 'costline', 'car_document', 'car_changelogs.car_changelog_items', 'checklist']);
+        $cars = Car::with(['user', 'costline', 'car_document.approvel_car_document', 'car_changelogs.car_changelog_items', 'checklist']);
         $users = User::where('role_id', 12)
             ->get();
         if ($user->role_id !== 1) {
@@ -109,7 +110,7 @@ class CarsController extends Controller
     public function saveImage($photo, $path, $name): String
     {
         try {
-            $imagename = time() . '_' . $name . $photo->getClientOriginalName();
+            $imagename = uniqid() . '_' . time() . '_' . $name . $photo->getClientOriginalName();
             $photo->move(public_path($path), $imagename);
             return $imagename;
         } catch (Exception $e) {
@@ -172,13 +173,12 @@ class CarsController extends Controller
             foreach ($archives as $archive) {
                 if ($request->hasFile($archive)) {
                     $document = $request->file($archive);
-                    $data[$archive] = time() . '_' . $document->getClientOriginalName();
+                    $data[$archive] = uniqid() . '_' . time() . '_' . $document->getClientOriginalName();
                     $document->move(public_path('documents/fleetcar/car_documents/'), $data[$archive]);
                 }
             }
-            // $data['car_id'] = $car->id;
             $carDocument = CarDocument::create($data);
-            return response()->json($carDocument);
+            return response()->json($carDocument, 200);
         } catch (Exception $e) {
             return response()->json($e->getMessage(), 500);
         }
@@ -188,22 +188,37 @@ class CarsController extends Controller
     {
         $data = $request->validated();
         $archives = ['ownership_card', 'technical_review', 'soat', 'insurance'];
-        foreach ($archives as $archive) {
-            if ($request->hasFile($archive)) {
-                $fileName = $carDocument->$archive;
-                if ($fileName) {
-                    $file_path = "documents/fleetcar/car_documents/$fileName";
-                    if (file_exists(public_path($file_path))) {
-                        unlink(public_path($file_path));
+        $user = Auth::user();
+        if ($user->role_id == 1) {
+            foreach ($archives as $archive) {
+                if ($request->hasFile($archive)) {
+                    $fileName = $carDocument->$archive;
+                    if ($fileName) {
+                        $file_path = "documents/fleetcar/car_documents/$fileName";
+                        if (file_exists(public_path($file_path))) {
+                            unlink(public_path($file_path));
+                        }
                     }
+                    $document = $request->file($archive);
+                    $data[$archive] = uniqid() . '_' . time() . '_' . $document->getClientOriginalName();
+                    $document->move(public_path('documents/fleetcar/car_documents/'), $data[$archive]);
                 }
-                $document = $request->file($archive);
-                $data[$archive] = time() . '_' . $document->getClientOriginalName();
-                $document->move(public_path('documents/fleetcar/car_documents/'), $data[$archive]);
             }
+            $carDocument->update($data);
+            $carDocument->load("approvel_car_document");
+            return response()->json($carDocument, 200);
+        } else {
+            foreach ($archives as $archive) {
+                if ($request->hasFile($archive)) {
+                    $document = $request->file($archive);
+                    $data[$archive] = uniqid() . '_' . time() . '_' . $document->getClientOriginalName();
+                    $document->move(public_path('documents/fleetcar/car_documents/'), $data[$archive]);
+                }
+            }
+            $data['car_document_id'] = $carDocument->id;
+            ApprovalCarDocument::create($data);
+            return response()->json([], 200);
         }
-        $carDocument->update($data);
-        return response()->json($carDocument);
     }
 
     public function destroyDocument(CarDocument $carDocument)
@@ -226,7 +241,7 @@ class CarsController extends Controller
         $data = $request->validated();
         if ($request->hasFile('invoice')) {
             $document = $request->file('invoice');
-            $data['invoice'] = time() . '_' . $document->getClientOriginalName();
+            $data['invoice'] = uniqid() . '_' . time() . '_' . $document->getClientOriginalName();
             $document->move(public_path('documents/fleetcar/invoices'), $data['invoice']);
         }
         $data['car_id'] = $car->id;
@@ -253,7 +268,7 @@ class CarsController extends Controller
                 unlink(public_path($file_path));
             }
             $document = $request->file('invoice');
-            $data['invoice'] = time() . '_' . $document->getClientOriginalName();
+            $data['invoice'] = uniqid() . '_' . time() . '_' . $document->getClientOriginalName();
             $document->move(public_path('documents/fleetcar/invoices'), $data['invoice']);
         }
         $carChangelog->update($data);
@@ -333,5 +348,81 @@ class CarsController extends Controller
         }
 
         return response()->json($images);
+    }
+
+
+    public function indexApprovelCarDocument()
+    {
+        $changes = ApprovalCarDocument::with([
+            'car_document:id,car_id',
+            'car_document.car:id,plate'
+        ])->get();
+        return Inertia::render("FleetCar/IndexApprovals", [
+            'change' => $changes
+        ]);
+    }
+
+    public function approveChanges($id)
+    {
+        $archives = ['ownership_card', 'technical_review', 'soat', 'insurance'];
+        try {
+            $approve_car = ApprovalCarDocument::find($id);
+            if ($approve_car) {
+                $car = CarDocument::find($approve_car->car_document_id);
+                foreach ($archives as $archive) {
+                    if ($car->$archive && $car->$archive !== $approve_car->$archive) {
+                        $fileName = $car->$archive;
+                        $file_path = "documents/fleetcar/car_documents/$fileName";
+                        if (file_exists(public_path($file_path))) {
+                            unlink(public_path($file_path));
+                        }
+                    }
+                }
+                $car->update($approve_car->toArray());
+
+                $approve_car->delete();
+                $pendingApprovals = ApprovalCarDocument::where('car_document_id', $car->id)->get();
+
+                foreach ($pendingApprovals as $approval) {
+                    foreach ($archives as $item) {
+                        if ($approval->$item) {
+                            $file_path = "documents/fleetcar/car_documents/{$approval->file_name}";
+                            if (file_exists(public_path($file_path))) {
+                                unlink(public_path($file_path));
+                            }
+                        }
+                    }
+                    $approval->delete();
+                }
+            }
+            $approve_car->delete();
+            return response()->json([], 200);
+        } catch (\Exception $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ], 500);
+        }
+    }
+
+    public function deleteChanges($id)
+    {
+        $archives = ['ownership_card', 'technical_review', 'soat', 'insurance'];
+        $changes = ApprovalCarDocument::find($id);
+        foreach ($archives as $item) {
+            if ($changes->$item) {
+                $file_path = "documents/fleetcar/car_documents/{$changes->$item}";
+                if (file_exists(public_path($file_path))) {
+                    unlink(public_path($file_path));
+                }
+            }
+        }
+        $changes->delete();
+        return response()->json([], 200);
+    }
+
+    public function approveAlarms()
+    {
+        $approval = CarDocument::whereHas('approvel_car_document')->get();
+        return response()->json($approval, 200);
     }
 }
